# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for dev.",
        "title": "dtle API Docs",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v2/tenant/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get tenant list.",
                "tags": [
                    "user"
                ],
                "operationId": "TenantListV2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TenantListResp"
                        }
                    }
                }
            }
        },
        "/v2/user/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create user.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "CreateUserV2",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserReqV2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserRespV2"
                        }
                    }
                }
            }
        },
        "/v2/user/current_user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get current user.",
                "tags": [
                    "user"
                ],
                "operationId": "GetCurrentUserV2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CurrentUserResp"
                        }
                    }
                }
            }
        },
        "/v2/user/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete user.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "DeleteUserV2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user name",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteUserRespV2"
                        }
                    }
                }
            }
        },
        "/v2/user/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get user list.",
                "tags": [
                    "user"
                ],
                "operationId": "UserListV2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filter user name",
                        "name": "filter_username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter tenant",
                        "name": "filter_tenant",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserListResp"
                        }
                    }
                }
            }
        },
        "/v2/user/list_action": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "list user action.",
                "tags": [
                    "user"
                ],
                "operationId": "ListActionV2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListActionRespV2"
                        }
                    }
                }
            }
        },
        "/v2/user/reset_password": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "reset user password.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "ResetPasswordV2",
                "parameters": [
                    {
                        "description": "reset user password",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetPasswordReqV2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResetPasswordRespV2"
                        }
                    }
                }
            }
        },
        "/v2/user/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update user info.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "UpdateUserV2",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserReqV2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserRespV2"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.User": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tenant": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.ButtonItem": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "text_cn": {
                    "type": "string"
                },
                "text_en": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateUserReqV2": {
            "type": "object",
            "required": [
                "pass_word",
                "role",
                "tenant",
                "username"
            ],
            "properties": {
                "pass_word": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tenant": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.CreateUserRespV2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.CurrentUserResp": {
            "type": "object",
            "properties": {
                "current_user": {
                    "$ref": "#/definitions/common.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.DeleteUserRespV2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ListActionRespV2": {
            "type": "object",
            "properties": {
                "authority": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MenuItem"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.MenuItem": {
            "type": "object",
            "properties": {
                "admin_only": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "menu_level": {
                    "type": "integer"
                },
                "menu_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ButtonItem"
                    }
                },
                "parent_id": {
                    "type": "integer"
                },
                "text_cn": {
                    "type": "string"
                },
                "text_en": {
                    "type": "string"
                }
            }
        },
...
        "models.ResetPasswordReqV2": {
            "type": "object",
            "required": [
                "current_user_password",
                "password",
                "tenant",
                "username"
            ],
            "properties": {
                "current_user_password": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "tenant": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ResetPasswordRespV2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.TenantListResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "tenant_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.UpdateUserReqV2": {
            "type": "object",
            "required": [
                "role",
                "tenant",
                "username"
            ],
            "properties": {
                "remark": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tenant": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UpdateUserRespV2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.UserListResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "user_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.User"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}